# docs/openapi/mnemonic-v1.yaml
openapi: 3.1.0
info:
  title: Mnemonic Admin REST API
  description: |
    Admin REST API for Mnemonic, a knowledge graph and tooling synchronization server.
    This API manages patterns, agents, and skills. It is consumed by curl or any HTTP
    client for data ingestion and administration.

    The MCP server (port 8081) is a separate protocol for Claude Code integration and
    is not documented in this OpenAPI specification.

    ## Architecture

    Mnemonic is a single Go binary that exposes two protocols:

    - **Admin REST API** (this spec) on port 8080: Read + Write via HTTP
    - **MCP Server** on port 8081: Read-only tools for Claude Code via MCP over Streamable HTTP

    ## Authentication

    MVP: no authentication. The server runs on localhost in a trusted environment.

    Post-MVP: authentication and authorization handled externally by Envoy and OPA.

    ## Error Handling

    All errors follow [RFC 7807 Problem Details](https://tools.ietf.org/html/rfc7807)
    with content type `application/problem+json`. Every response includes an
    `X-Request-ID` correlation header.

    ## Pagination

    List endpoints use cursor-based pagination. Cursors are opaque base64-encoded
    strings. The `has_more` field indicates whether additional pages exist.

    ## Request Correlation

    Every response includes an `X-Request-ID` header. If the client provides this
    header on the request, the server echoes it back. Otherwise the server generates
    a UUID v4.
...
servers:
...
components:
  parameters:
...
    skillNameParam:
      name: name
      in: path
      required: true
      description: Skill unique identifier (lowercase letters, numbers, and hyphens; no leading/trailing hyphens, no consecutive hyphens)
      schema:
        type: string
        pattern: ^[a-z]([a-z0-9](-[a-z0-9])*)*$
        maxLength: 64
      example: mnemonic-sync
    skillFileNameParam:
      name: filename
      in: path
      required: true
      description: File name within the skill directory (e.g., "extract.py", "REFERENCE.md")
      schema:
        type: string
        pattern: ^[a-zA-Z0-9][a-zA-Z0-9._-]*$
        maxLength: 255
      example: extract.py
...
  headers:
    X-Request-ID:
      description: Correlation ID for request tracing. Echoed from request or generated by server.
      schema:
        type: string
        format: uuid
      example: 550e8400-e29b-41d4-a716-446655440000
    Location:
      description: URI of the newly created resource
      schema:
        type: string
        format: uri-reference
  schemas:
...
    SkillFile:
      type: object
      description: A file (script, reference, or asset) belonging to a skill
      required:
        - filename
        - content_type
        - size
        - created_at
        - updated_at
      properties:
        filename:
          type: string
          pattern: ^[a-zA-Z0-9][a-zA-Z0-9._-]*$
          maxLength: 255
          description: File name within the skill directory (e.g., "extract.py", "REFERENCE.md")
        content_type:
          type: string
          maxLength: 128
          description: MIME type (e.g., "text/x-python", "text/markdown", "image/png")
        content:
          type: string
          description: File content. Text files as UTF-8 string, binary files as base64-encoded string.
        encoding:
          type: string
          enum:
            - utf-8
            - base64
          default: utf-8
          description: Content encoding. Use base64 for binary assets.
        size:
          type: integer
          readOnly: true
          description: File size in bytes
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
    SkillFileCreate:
      type: object
      description: Request body for uploading a file to a skill
      required:
        - filename
        - content_type
        - content
      properties:
        filename:
          type: string
          pattern: ^[a-zA-Z0-9][a-zA-Z0-9._-]*$
          maxLength: 255
        content_type:
          type: string
          maxLength: 128
        content:
          type: string
          description: File content (UTF-8 text or base64-encoded binary)
        encoding:
          type: string
          enum:
            - utf-8
            - base64
          default: utf-8
    SkillFileUpdate:
      type: object
      description: Request body for replacing a file in a skill
      required:
        - content_type
        - content
      properties:
        content_type:
          type: string
          maxLength: 128
        content:
          type: string
        encoding:
          type: string
          enum:
            - utf-8
            - base64
          default: utf-8
    SkillFileSummary:
      type: object
      description: File metadata without content (for list responses)
      required:
        - filename
        - content_type
        - size
        - created_at
        - updated_at
      properties:
        filename:
          type: string
        content_type:
          type: string
        size:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SkillFileList:
      type: object
      description: List of file summaries for a skill child collection
      required:
        - data
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/SkillFileSummary"

    # ==================== Health/Operations Schemas ====================
...
  responses:
    BadRequest:
      description: Validation error or malformed request
      headers:
        X-Request-ID:
          $ref: "#/components/headers/X-Request-ID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            type: https://mnemonic.example.com/problems/validation-error
            title: Validation Error
            status: 400
            detail: The request body contains invalid fields
            instance: /v1/api/agents
            traceId: 550e8400-e29b-41d4-a716-446655440000
            errors:
              - field: name
                code: INVALID_FORMAT
                message: "Name must match pattern ^[a-z]([a-z0-9](-[a-z0-9])*)*$"
              - field: system_prompt
                code: REQUIRED
                message: System prompt is required
    NotFound:
      description: Resource not found
      headers:
        X-Request-ID:
          $ref: "#/components/headers/X-Request-ID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            type: https://mnemonic.example.com/problems/not-found
            title: Not Found
            status: 404
            detail: "Agent with name 'unknown-agent' not found"
            instance: /v1/api/agents/unknown-agent
            traceId: 550e8400-e29b-41d4-a716-446655440000
    Conflict:
      description: Resource conflict (e.g., duplicate name)
      headers:
        X-Request-ID:
          $ref: "#/components/headers/X-Request-ID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            type: https://mnemonic.example.com/problems/conflict
            title: Conflict
            status: 409
            detail: "Agent with name 'go-software-engineer' already exists"
            traceId: 550e8400-e29b-41d4-a716-446655440000
    LimitExceeded:
      description: Collection limit exceeded (e.g., max files per skill)
      headers:
        X-Request-ID:
          $ref: "#/components/headers/X-Request-ID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            type: https://mnemonic.example.com/problems/limit-exceeded
            title: Limit Exceeded
            status: 422
            detail: "Skill 'my-skill' already has the maximum of 20 script files"
            traceId: 550e8400-e29b-41d4-a716-446655440000
    ContentTooLarge:
      description: Request payload exceeds the size limit for this resource type
      headers:
        X-Request-ID:
          $ref: "#/components/headers/X-Request-ID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            type: https://mnemonic.example.com/problems/content-too-large
            title: Content Too Large
            status: 413
            detail: "Script file exceeds maximum size of 1048576 bytes"
            traceId: 550e8400-e29b-41d4-a716-446655440000
    InternalError:
      description: Internal server error
      headers:
        X-Request-ID:
          $ref: "#/components/headers/X-Request-ID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            type: https://mnemonic.example.com/problems/internal-error
            title: Internal Error
            status: 500
            detail: An unexpected error occurred
            traceId: 550e8400-e29b-41d4-a716-446655440000
...
paths:
...
  /v1/api/skills/{name}/scripts:
    parameters:
      - $ref: "#/components/parameters/skillNameParam"

    get:
      tags:
        - Skill Files
      summary: List scripts for a skill
      description: |
        Returns metadata for all script files belonging to the skill.
        Content is not included; fetch individual files to retrieve content.
      operationId: listSkillScripts
      responses:
        "200":
          description: List of script file summaries
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillFileList"
              example:
                data:
                  - filename: extract.py
                    content_type: text/x-python
                    size: 2048
                    created_at: "2026-02-15T10:00:00Z"
                    updated_at: "2026-02-15T10:00:00Z"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      tags:
        - Skill Files
      summary: Upload script to a skill
      description: |
        Uploads a new script file to the skill. The filename must be unique
        within the skill's scripts collection. Maximum file size: 1MB (1048576 bytes).
        Maximum 20 script files per skill.
      operationId: createSkillScript
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillFileCreate"
            example:
              filename: extract.py
              content_type: text/x-python
              content: "#!/usr/bin/env python3\nimport sys\n\ndef extract():\n    pass\n"
              encoding: utf-8
      responses:
        "201":
          description: Script uploaded
          headers:
            Location:
              $ref: "#/components/headers/Location"
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillFile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          $ref: "#/components/responses/ContentTooLarge"
        "422":
          $ref: "#/components/responses/LimitExceeded"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/api/skills/{name}/scripts/{filename}:
    parameters:
      - $ref: "#/components/parameters/skillNameParam"
      - $ref: "#/components/parameters/skillFileNameParam"

    get:
      tags:
        - Skill Files
      summary: Get script by filename
      description: Returns the script file with content included.
      operationId: getSkillScript
      responses:
        "200":
          description: Script file with content
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillFile"
              example:
                filename: extract.py
                content_type: text/x-python
                content: "#!/usr/bin/env python3\nimport sys\n\ndef extract():\n    pass\n"
                encoding: utf-8
                size: 54
                created_at: "2026-02-15T10:00:00Z"
                updated_at: "2026-02-15T10:00:00Z"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      tags:
        - Skill Files
      summary: Replace script
      description: Replaces the content of an existing script file.
      operationId: updateSkillScript
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillFileUpdate"
      responses:
        "200":
          description: Script replaced
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillFile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "413":
          $ref: "#/components/responses/ContentTooLarge"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      tags:
        - Skill Files
      summary: Delete script
      operationId: deleteSkillScript
      responses:
        "204":
          description: Script deleted
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # --- References ---
  /v1/api/skills/{name}/references:
    parameters:
      - $ref: "#/components/parameters/skillNameParam"

    get:
      tags:
        - Skill Files
      summary: List references for a skill
      description: |
        Returns metadata for all reference files belonging to the skill.
        Content is not included; fetch individual files to retrieve content.
      operationId: listSkillReferences
      responses:
        "200":
          description: List of reference file summaries
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillFileList"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      tags:
        - Skill Files
      summary: Upload reference to a skill
      description: |
        Uploads a new reference file to the skill. Maximum file size: 1MB (1048576 bytes).
        Maximum 50 reference files per skill.
      operationId: createSkillReference
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillFileCreate"
            example:
              filename: REFERENCE.md
              content_type: text/markdown
              content: "# Additional Reference\n\nThis document provides..."
              encoding: utf-8
      responses:
        "201":
          description: Reference uploaded
          headers:
            Location:
              $ref: "#/components/headers/Location"
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillFile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          $ref: "#/components/responses/ContentTooLarge"
        "422":
          $ref: "#/components/responses/LimitExceeded"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/api/skills/{name}/references/{filename}:
    parameters:
      - $ref: "#/components/parameters/skillNameParam"
      - $ref: "#/components/parameters/skillFileNameParam"

    get:
      tags:
        - Skill Files
      summary: Get reference by filename
      description: Returns the reference file with content included.
      operationId: getSkillReference
      responses:
        "200":
          description: Reference file with content
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillFile"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      tags:
        - Skill Files
      summary: Replace reference
      description: Replaces the content of an existing reference file.
      operationId: updateSkillReference
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillFileUpdate"
      responses:
        "200":
          description: Reference replaced
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillFile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "413":
          $ref: "#/components/responses/ContentTooLarge"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      tags:
        - Skill Files
      summary: Delete reference
      operationId: deleteSkillReference
      responses:
        "204":
          description: Reference deleted
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # --- Assets ---
  /v1/api/skills/{name}/assets:
    parameters:
      - $ref: "#/components/parameters/skillNameParam"

    get:
      tags:
        - Skill Files
      summary: List assets for a skill
      description: |
        Returns metadata for all asset files belonging to the skill.
        Content is not included; fetch individual files to retrieve content.
      operationId: listSkillAssets
      responses:
        "200":
          description: List of asset file summaries
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillFileList"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      tags:
        - Skill Files
      summary: Upload asset to a skill
      description: |
        Uploads a new asset file to the skill. Maximum file size: 5MB (5242880 bytes).
        Maximum 50 asset files per skill. Use base64 encoding for binary content.
      operationId: createSkillAsset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillFileCreate"
            example:
              filename: template.json
              content_type: application/json
              content: "{\"key\": \"value\"}"
              encoding: utf-8
      responses:
        "201":
          description: Asset uploaded
          headers:
            Location:
              $ref: "#/components/headers/Location"
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillFile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          $ref: "#/components/responses/ContentTooLarge"
        "422":
          $ref: "#/components/responses/LimitExceeded"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/api/skills/{name}/assets/{filename}:
    parameters:
      - $ref: "#/components/parameters/skillNameParam"
      - $ref: "#/components/parameters/skillFileNameParam"

    get:
      tags:
        - Skill Files
      summary: Get asset by filename
      description: Returns the asset file with content included.
      operationId: getSkillAsset
      responses:
        "200":
          description: Asset file with content
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillFile"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      tags:
        - Skill Files
      summary: Replace asset
      description: Replaces the content of an existing asset file.
      operationId: updateSkillAsset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillFileUpdate"
      responses:
        "200":
          description: Asset replaced
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillFile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "413":
          $ref: "#/components/responses/ContentTooLarge"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      tags:
        - Skill Files
      summary: Delete asset
      operationId: deleteSkillAsset
      responses:
        "204":
          description: Asset deleted
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ==================== Operations Endpoints ====================
...